requirements_docker.txt
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.1.1
cycler==0.12.1
filelock==3.16.1
flatbuffers==25.2.10
fonttools==4.57.0
humanfriendly==10.0
idna==3.10
importlib_resources==6.4.5
Jinja2==3.1.6
kiwisolver==1.4.7
MarkupSafe==2.1.5
matplotlib==3.7.5
mpmath==1.3.0
networkx==3.1
numpy==1.24.3
onnxruntime==1.16.3
opencv-python==4.8.0.76
packaging==24.2
pandas==2.0.3
pillow==10.4.0
protobuf==5.29.4
psutil==7.0.0
py-cpuinfo==9.0.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
scipy==1.10.1
seaborn==0.13.2
six==1.17.0
sympy==1.13.3
thop==0.1.1.post2209072238
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tqdm==4.67.1
typing_extensions==4.13.2
tzdata==2025.2
ultralytics==8.0.196
urllib3==2.2.3
zipp==3.20.2
# Added for containerization
streamlit==1.28.1
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
